Renamed layout components to match others
